services/worker/requirements.txt
# Queue and database
redis[hiredis]>=5.0.0
asyncpg>=0.29.0

# Microsoft Graph API
httpx>=0.25.0
msal>=1.25.0
cryptography>=42.0.0

# Text extraction
PyMuPDF>=1.23.0
python-docx>=0.8.11
openpyxl>=3.1.0
python-pptx>=0.6.21

# OCR
pytesseract>=0.3.10

# Image processing
Pillow>=10.0.0
pillow-heif>=0.13.0

# AI providers
anthropic>=0.37.0
openai>=1.6.0
# Gemini uses httpx (already listed above) to call Vertex AI REST API directly

# HTML sanitization
bleach[css]>=6.0.0

# OneNote extraction
pyOneNote>=0.0.2
python-magic>=0.4.27

# Utilities
pydantic>=2.5.0
python-json-logger>=2.0.0
jinja2>=3.1.0
pyyaml>=6.0.0
